refactor: rename torch stubs
