+++ _torch_stub.py
