+++ torchvision_stub/__init__.py
