+++ torchvision_stub/models.py
